+++ components/collector/src/source_collectors/__init__.py
@@ -81,6 +81,7 @@ from .jira.test_cases import JiraTestCases
 from .jira.user_story_points import JiraUserStoryPoints
 from .jira.velocity import JiraVelocity
 
+from .jmeter_csv.performancetest_duration import JMeterCSVPerformanceTestDuration
 from .jmeter_csv.slow_transactions import JMeterCSVSlowTransactions
 from .jmeter_csv.tests import JMeterCSVTests