+++ docs/src/changelog.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 <!-- The line "## <square-bracket>Unreleased</square-bracket>" is replaced by the release/release.py script with the new release version and release date. -->
 
-## v3.30.0-rc.2 - 2021-12-09
+## [Unreleased]
 
 ### Fixed
 
@@ -17,7 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ### Added
 
-- Support JMeter CSV output as source for the 'slow transactions' and 'tests' metrics. Closes [#2966](https://github.com/ICTU/quality-time/issues/2966) and [#2967](https://github.com/ICTU/quality-time/issues/2967).
+- Support JMeter CSV output as source for the 'performance test duration', 'slow transactions' and 'tests' metrics. Closes [#2965](https://github.com/ICTU/quality-time/issues/2965), [#2966](https://github.com/ICTU/quality-time/issues/2966) and [#2967](https://github.com/ICTU/quality-time/issues/2967).
 
 ## v3.29.0 - 2021-12-03